requirements.txt
absl-py==0.13.0
appdirs==1.4.4
asgiref==3.4.1
astor==0.8.1
audioread==2.1.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==5.0.9
Django==3.2.5
djangorestframework==3.12.4
gast==0.2.2
google-pasta==0.2.0
grpcio==1.38.1
h5py==2.10.0
hparams==0.3.0
idna==2.10
importlib-metadata==4.6.1
inflect==5.3.0
jamo==0.4.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.2
numba==0.53.1
numpy==1.18.5
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.1
pooch==1.4.0
protobuf==3.17.3
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.1
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.61.2
typeguard==2.12.1
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
